requirements.txt
Django>=4.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=3.13.0
psycopg2>=2.9.5
pillow>=9.4.0